Fix issue where unset env vars break makefile
We construct image digests using bash string processing, which requires
the image to be set in an environment varible. However, in the makefile,
there are default values for variables that don't get propagated to the
environment, meaning make install and make build fail when those
default values are used.

Instead, we explicitly set the environment before constructing the
digest.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>