Update images in templates even if state is unmanaged
If a template is in an unmanaged state but still uses a default image
(i.e. one from the same image repository as the new one), update only
the image in the unmanaged template to ensure CVE fixes still get
applied to templates that change e.g. default resource limits/requests.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>